added tonapi tests

+++ Windetta.Wallet/src/Application/Services/IUserWalletService.cs
@@ -6,7 +6,7 @@ namespace Windetta.Wallet.Application.Services;
 public interface IUserWalletService : IScopedService
 {
     Task<WalletInfoDto> GetWalletInfoAsync(Guid userId);
-    Task HoldBalanceAsync(Guid userId, int nanotons);
+    Task HoldBalanceAsync(Guid userId, long nanotons);
     Task UnHoldBalanceAsync(Guid userId);
     Task<WalletInfoDto> CreateWalletAsync(Guid userId);
     Task TransferAsync(Guid userId, long nanotons, TonAddress destinationAddress);
